GRAPHICS: Add base console for KotOR games
