+++ src/engines/kotor/console.cpp
@@ -22,198 +22,19 @@
  *  Star Wars: Knights of the Old Republic (debug) console.
  */
 
-#include <algorithm>
-
-#include <boost/bind.hpp>
-
-#include "src/common/ustring.h"
-#include "src/common/util.h"
-#include "src/common/filepath.h"
-#include "src/common/filelist.h"
-#include "src/common/configman.h"
-#include "src/common/strutil.h"
-
-#include "src/aurora/resman.h"
-
-#include "src/graphics/aurora/types.h"
-
-#include "src/engines/kotorbase/room.h"
-#include "src/engines/kotorbase/module.h"
-#include "src/engines/kotorbase/area.h"
-
 #include "src/engines/kotor/console.h"
 #include "src/engines/kotor/kotor.h"
-#include "src/engines/kotor/game.h"
-#include "src/engines/kotor/creature.h"
 
 namespace Engines {
 
 namespace KotOR {
 
-Console::Console(KotOREngine &engine) :
-	::Engines::Console(engine, Graphics::Aurora::kSystemFontMono, 13),
-	_engine(&engine), _maxSizeMusic(0) {
-
-	registerCommand("exitmodule"          , boost::bind(&Console::cmdExitModule          , this, _1),
-			"Usage: exitmodule\nExit the module, returning to the main menu");
-	registerCommand("listmodules"         , boost::bind(&Console::cmdListModules         , this, _1),
-			"Usage: listmodules\nList all modules");
-	registerCommand("loadmodule"          , boost::bind(&Console::cmdLoadModule          , this, _1),
-			"Usage: loadmodule <module>\nLoad and enter the specified module");
-	registerCommand("listmusic"           , boost::bind(&Console::cmdListMusic           , this, _1),
-			"Usage: listmusic\nList all available music resources");
-	registerCommand("stopmusic"           , boost::bind(&Console::cmdStopMusic           , this, _1),
-			"Usage: stopmusic\nStop the currently playing music resource");
-	registerCommand("playmusic"           , boost::bind(&Console::cmdPlayMusic           , this, _1),
-			"Usage: playmusic [<music>]\nPlay the specified music resource. "
-			"If none was specified, play the default area music.");
-	registerCommand("flycam"              , boost::bind(&Console::cmdFlyCam              , this, _1),
-			"Usage: flycam\nToggle free roam camera mode");
-	registerCommand("showwalkmesh"        , boost::bind(&Console::cmdShowWalkmesh        , this, _1),
-			"Usage: showwalkmesh\nToggle walkmesh display");
-	registerCommand("showtriggers"        , boost::bind(&Console::cmdShowTriggers        , this, _1),
-			"Usage: showtriggers\nToggle triggers display");
-	registerCommand("getpcroom"           , boost::bind(&Console::cmdGetPCRoom           , this, _1),
-			"Usage: getpcroom\nGet a room PC is in");
-	registerCommand("listroomsvisiblefrom", boost::bind(&Console::cmdListRoomsVisibleFrom, this, _1),
-			"Usage: listroomsvisiblefrom <room>\nList rooms that are visible from the specified room");
-	registerCommand("playanim"            , boost::bind(&Console::cmdPlayAnim            , this, _1),
-			"Usage: playanim <base> [<head>]\nPlay the specified animations on the active object");
-	registerCommand("additem"             , boost::bind(&Console::cmdAddItem             , this, _1),
-			"Usage: additem <item> [<count>]\nAdd the specified item to the active object");
-
+Console::Console(KotOREngine &engine) : KotORBase::Console(engine) {
 }
 
-Console::~Console() {
-}
-
-void Console::updateCaches() {
-	::Engines::Console::updateCaches();
-
-	updateModules();
-	updateMusic();
-}
-
-void Console::updateModules() {
-	_modules = _engine->getGame().getModules();
-
-	setArguments("loadmodule", _modules);
-}
-
-void Console::updateMusic() {
-	_music.clear();
-	_maxSizeMusic = 0;
-
-	const Common::FileList music(Common::FilePath::findSubDirectory(ResMan.getDataBase(), "streammusic", true));
-
-	for (Common::FileList::const_iterator m = music.begin(); m != music.end(); ++m) {
-		if (!Common::FilePath::getExtension(*m).equalsIgnoreCase(".wav"))
-			continue;
-
-		_music.push_back(Common::FilePath::getStem(*m));
-
-		_maxSizeMusic = MAX(_maxSizeMusic, _music.back().size());
-	}
-
-	std::sort(_music.begin(), _music.end(), Common::UString::iless());
-	setArguments("playmusic", _music);
-}
-
-void Console::cmdExitModule(const CommandLine &UNUSED(cl)) {
-	hide();
-	_engine->getGame().getModule().exit();
-}
-
-void Console::cmdListModules(const CommandLine &UNUSED(cl)) {
-	updateModules();
-
-	for (std::vector<Common::UString>::iterator m = _modules.begin(); m != _modules.end(); ++m)
-		printf("%s (\"%s\")", m->c_str(), KotORBase::Module::getName(*m, "KOTOR_moduleDir").c_str());
-}
-
-void Console::cmdLoadModule(const CommandLine &cl) {
-	if (cl.args.empty()) {
-		printCommandHelp(cl.cmd);
-		return;
-	}
-
-	if (_engine->getGame().hasModule(cl.args)) {
-		hide();
-		_engine->getGame().getModule().load(cl.args);
-		return;
-	}
-
-	printf("No such module \"%s\"", cl.args.c_str());
-}
-
-void Console::cmdListMusic(const CommandLine &UNUSED(cl)) {
-	updateMusic();
-	printList(_music, _maxSizeMusic);
-}
-
-void Console::cmdStopMusic(const CommandLine &UNUSED(cl)) {
-	_engine->getGame().stopMusic();
-}
-
-void Console::cmdPlayMusic(const CommandLine &cl) {
-	_engine->getGame().playMusic(cl.args);
-}
-
-void Console::cmdFlyCam(const CommandLine &UNUSED(cl)) {
-	_engine->getGame().getModule().toggleFlyCamera();
-}
-
-void Console::cmdShowWalkmesh(const CommandLine &UNUSED(cl)) {
-	_engine->getGame().getModule().toggleWalkmesh();
-}
-
-void Console::cmdShowTriggers(const CommandLine &UNUSED(cl)) {
-	_engine->getGame().getModule().toggleTriggers();
-}
-
-void Console::cmdGetPCRoom(const CommandLine &UNUSED(cl)) {
-	const KotORBase::Room *room = _engine->getGame().getModule().getPC()->getRoom();
-	printf("%s", room->getResRef().c_str());
-}
-
-void Console::cmdListRoomsVisibleFrom(const CommandLine &cl) {
-	if (cl.args.empty()) {
-		printCommandHelp(cl.cmd);
-		return;
-	}
-
-	const std::vector<Common::UString> &rooms = _engine->getGame().getModule().getCurrentArea()->getRoomsVisibleFrom(cl.args);
-	for (std::vector<Common::UString>::const_iterator r = rooms.begin();
-			r != rooms.end(); ++r) {
-		printf("%s", r->c_str());
-	}
-}
-
-void Console::cmdPlayAnim(const CommandLine &cl) {
-	if (cl.args.empty()) {
-		printCommandHelp(cl.cmd);
-		return;
-	}
-	std::vector<Common::UString> anims;
-	Common::UString::split(cl.args, ' ', anims);
-	size_t animCount = anims.size();
-	_engine->getGame().getModule().playAnimationOnActiveObject(anims[0],
-			animCount >= 2 ? anims[1] : "");
-}
-
-void Console::cmdAddItem(const CommandLine &cl) {
-	if (cl.args.empty()) {
-		printCommandHelp(cl.cmd);
-		return;
-	}
-
-	std::vector<Common::UString> args;
-	Common::UString::split(cl.args, ' ', args);
-	int count = 1;
-	if (args.size() >= 2)
-		Common::parseString(args[1], count);
-
-	_engine->getGame().getModule().addItemToActiveObject(args[0], count);
+const Common::UString &Console::getModuleDirOptionName() const {
+	static Common::UString name("KOTOR_moduleDir");
+	return name;
 }
 
 } // End of namespace KotOR
